requirements.txt
streamlit
pandas
PyPDF
langchain
textract
networkx
numpy
